+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Added folding and custom highlighting to backlinks window.
+- When the title of a note is changed, the title will automatically be added to note's aliases in the frontmatter on save.
 
 ### Fixed